+++ crayg/PineHoleCameraModel.h
@@ -8,9 +8,11 @@
 
 #include "scene/Camera.h"
 #include "foundation/Ray.h"
+#include "CameraModel.h"
+
 // todo add tests
 // todo proper virtual class
-class PineHoleCameraModel {
+class PineHoleCameraModel : public CameraModel{
 private:
     Camera &camera;
     float imageRatio;
@@ -24,7 +26,7 @@ private:
 public:
     PineHoleCameraModel(Camera &camera, int imageWidth, int imageHeight);
 
-    Ray createPrimaryRay(float x, float y);
+    Ray createPrimaryRay(float x, float y) override;
 };